changes made by @sanjay270899 now follow the file-structure (N.B.:-@sanjay270899 header and footer will now replace @anushbhatia 's header and footer)
