+++ src/components/Footer/Footer.js
@@ -1,19 +1,12 @@
-import React from "react";
-import "./Footer.css";
-import { Fade } from "react-reveal";
-import emoji from "react-easy-emoji";
-/* eslint-disable jsx-a11y/accessible-emoji */
+import React from 'react';
+import styles from './Footer.module.css';
 
-export default function Footer(props) {
-  return (
-    //py-2 for padding
-    <div className="footer-div py-2">
-      <Fade>
-        <p className="footer-text">
-        {emoji("Made with ❤️ in India for the students of the world.")} 
-        </p>
-        {/* <ToggleSwitch theme={props.theme} onToggle={props.onToggle}/> */}
-      </Fade>
-    </div>
-  );
+function Footer() {
+    return (
+        <footer className={styles.Footer}>
+            <div>Made with ❤️ in India for the students of the world.</div>
+        </footer>
+    )
 }
+
+export default Footer
