--- src/pages/HomePage/Footer/Footer.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import styles from './Footer.module.css';
-
-function Footer() {
-    return (
-        <footer className={styles.Footer}>
-            <div>Made with ❤️ in India for the students of the world.</div>
-        </footer>
-    )
-}
-
-export default Footer
